backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiofiles==24.1.0
    # via -r requirements.in
alembic==1.16.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
bcrypt==4.3.0
    # via passlib
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via cryptography
click==8.2.1
    # via uvicorn
cryptography==45.0.4
    # via python-jose
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via -r requirements.in
fastapi==0.115.12
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
jiter==0.10.0
    # via openai
lxml==5.4.0
    # via
    #   -r requirements.in
    #   python-docx
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
openai==1.86.0
    # via -r requirements.in
passlib==1.7.4
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via -r requirements.in
pyjwt==2.10.1
    # via -r requirements.in
pypdf2==3.0.1
    # via -r requirements.in
python-docx==1.1.2
    # via -r requirements.in
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-jose==3.5.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via uvicorn
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via
    #   -r requirements.in
    #   alembic
starlette==0.46.2
    # via fastapi
tqdm==4.67.1
    # via openai
typing-extensions==4.14.0
    # via
    #   alembic
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.34.3
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
